Implemented number scraping via UI
Tied it all together. Right now though the UI freezes up, so I should probably make it asynchronous.

Also I changed up the DateFetcher so that the user no longer has to know and specify whether to start incrementing dates by 3s or 4s, nor do they have to select a start date that is a drawing date. The DateFetcher will now automatically start from the soonest Tuesday/Friday after the start date and begin incrementing by 3 or 4 depending on the drawing day. This will work better with the way the UI is set up, in that any date (after 10/31/17) can be selected via the DatePickers.